update: moves account methods to directory called 'Accounts' for modularity
